backend/requirements.txt
# ==============================================================================
# Aditus Backend Dependencies
# ==============================================================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - Async PostgreSQL with SQLAlchemy
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# AI/LLM
google-generativeai==0.8.6

# Web Scraping & Parsing
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
trafilatura==1.6.1

# PDF Generation (ReportLab for simple, reliable PDF generation)
reportlab==4.0.9
pydantic==2.5.0
Pillow==10.1.0

# Background Tasks
python-rq==1.14.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Gmail API & OAuth2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0

# HTTP utilities
requests==2.31.0
urllib3==2.1.0

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
python-logging-loki==0.3.2
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
ruff==0.1.8

# Utilities
python-dotenv==1.0.0
click==8.1.7
